tg_udarenia/requirements.txt

В моем коде используются следующие сторонние библиотеки:

1.telebot: Предоставляет API для работы с Telegram Bot API, используется для взаимодействия с Telegram и отправки сообщений, стикеров и других элементов пользователю.

2.sqlalchemy: SQL-текстовая библиотека и ORM для языка программирования Python. Используется для взаимодействия с базой данных SQLite, создания сессий и выполнения запросов.

3.sqlalchemy_serializer: Расширяет SQLAlchemy, добавляя возможность автоматической сериализации объектов базы данных в JSON.

4.unittest: Встроенная библиотека для написания и запуска тестов. Ваши тесты используют unittest для создания и запуска тестовых сценариев.